Add HotelDetail component and seperate App.js into Homescreen.js
Refactored the codebase by separating the logic from App.js into HomeScreen.js. This restructuring improves code organization and maintainability. Additionally, the HotelDetail component was added to the HomeScreen, allowing for the display of hotel images and details. The hotel data used is still dummy data stored in the Redux store.